+++ src/components/layer-controller/LayerController.js
@@ -10,6 +10,8 @@ import Tab from '@material-ui/core/Tab';
 
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import VisibilityIcon from '@material-ui/icons/Visibility';
+import VisibilityOffIcon from '@material-ui/icons/VisibilityOff';
 
 import LayerOptions from './LayerOptions';
 import VolumeOptions from './VolumeOptions';
@@ -111,6 +113,13 @@ export default function LayerController(props) {
       .filter(field => typeof firstSelection[field] === 'number')
       .reduce((o, key) => ({ ...o, [key]: firstSelection[key] }), {}),
   );
+
+  function setVisible(v) {
+    const newChannels = [...channels];
+    // eslint-disable-next-line no-param-reassign,no-return-assign
+    newChannels.forEach(ch => (ch.visible = v));
+    handleLayerChange({ ...layer, channels: newChannels });
+  }
 
   function setColormap(v) {
     handleLayerChange({ ...layer, colormap: v });
@@ -329,6 +338,8 @@ export default function LayerController(props) {
     setAreLayerChannelsLoading(newAreLayerChannelsLoading);
   };
 
+  const visible = channels.some(ch => ch.visible);
+  const Visibility = visible ? VisibilityIcon : VisibilityOffIcon;
   return (
     <ExpansionPanel
       className={controllerSectionClasses.root}
@@ -341,7 +352,21 @@ export default function LayerController(props) {
         aria-controls={`layer-${name}-controls`}
       >
         <Grid container direction="column" m={1} justify="center">
-          <OverflowEllipsisGrid item>{name}</OverflowEllipsisGrid>
+          <OverflowEllipsisGrid item>
+            <Button
+              onClick={(e) => {
+                // Needed to prevent affecting the expansion panel from changing
+                e.stopPropagation();
+                setVisible(!visible);
+              }}
+              style={{
+                marginRight: 8, marginBottom: 2, padding: 0, minWidth: 0,
+              }}
+            >
+              <Visibility />
+            </Button>
+            {name}
+          </OverflowEllipsisGrid>
           {!isExpanded && !use3D && (
             <Grid container direction="row" alignItems="center" justify="center">
               <Grid item xs={6}>