+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## In Progress
 
 ### Added
+
+- Global visibility button next to name in layer controller.
+
+### Changed
+- Fix bitmask remove button style.
+
+## [1.1.10](https://www.npmjs.com/package/vitessce/v/1.1.9) - 2021-05-19
+
+### Added
 - Add support for bitmasks to `Spatial` component and raster schema.
 - Worker pool for processing heatmap tiles.